+++ mcs/class/System/Makefile
@@ -86,6 +86,10 @@ ifdef ONLY_APPLETLS
 LIB_MCS_FLAGS += -d:ONLY_APPLETLS
 endif
 
+ifdef MONO_FEATURE_APPLE_X509
+LIB_MCS_FLAGS += -d:MONO_FEATURE_APPLE_X509
+endif
+
 ifndef PROFILE_DISABLE_BTLS
 ifdef HAVE_BTLS
