+++ mcs/class/System/monotouch_watch_System.dll.exclude.sources
@@ -1,4 +1,2 @@
 #include common_networking.sources
 #include appletls.sources
-Mono.AppleTls/MonoCertificatePal.Mobile.cs
-Mono.AppleTls/SafeHandles.cs
